fix(deployer): Add warning for invalid configuration (#6682)
## Description

While we have warnings in
https://github.com/mastra-ai/mastra/blob/main/packages/deployer/src/build/babel/remove-all-options-except.ts
these are sometimes not triggered because the bundling fails before that
step. Thus in this PR a new babel check is added inside `analyzeBundle`
that should run soon enough to catch everything.

This PR is on the cautious side so that's why it's a warning. I could
see us turning this into an error at a later point if we don't receive
any complaints/issues for this. If we think this is safe to make it an
error that's also fine for me.

## Related Issue(s)

Fixes COR-85
Fixes https://github.com/mastra-ai/mastra/issues/6621

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [x] I have added tests that prove my fix is effective or that my
feature works